Replace smart contract calls (#57)
* Removed old method of handling signing of auth type

* Update smart contract calls for auth

1. Added new `setClientIds` method
2. Replaced existing individual auth methods with `setClientIds`

* Modified the signerMakeTx function
