+++ src/pages/AppConfigure.vue
@@ -483,11 +483,10 @@ import {
   deleteApp as deleteAppApi,
   deleteCred,
 } from "@/services/app-config.service";
-import { signerMakeTx } from "../utils/signer";
-import getEnvApi from "../services/get-env-api";
-import { getAddress } from "../utils/get-address";
-import FullScreenLoader from "../components/FullScreenLoader.vue";
-import { decrypt } from "../utils/cryptoUtils";
+import signerMakeTx from "@/utils/signerMakeTx";
+import { getAddress } from "@/utils/get-address";
+import FullScreenLoader from "@/components/FullScreenLoader.vue";
+import { decrypt } from "@/utils/cryptoUtils";
 import useArcanaAuth from "@/use/arcanaAuth";
 
 export default {
@@ -675,109 +674,36 @@ export default {
 
     async function makeTx() {
       const config = { ...store.getters[env.value + "/config"] };
-      const ssoClients = [
-        {
-          type: "google",
-          method: "setGoogleClientId",
-        },
-        {
-          type: "github",
-          method: "setGithubClientId",
-        },
-        {
-          type: "twitter",
-          method: "setTwitterClientId",
-        },
-        {
-          type: "twitch",
-          method: "setTwitchClientId",
-        },
-        {
-          type: "reddit",
-          method: "setRedditClientId",
-        },
-        {
-          type: "discord",
-          method: "setDiscordClientId",
-        },
+
+      if (store.getters.appName !== previousConfig.appName) {
+        loadingMessage.value = `Updating app name in smart contract...`;
+        await signerMakeTx("setAppName", [store.getters.appName]);
+      }
+
+      const authDetails = [
+        ...store.getters[store.getters.env + "/authDetails"],
       ];
-      for (let ssoClient of ssoClients) {
-        try {
-          const clientId = store.getters[store.getters.env + "/authDetails"]
-            .find((el) => el.verifier === ssoClient.type)
-            ?.clientId?.trim();
+      const authSignerMakeTxValue = [[], []];
 
-          const previousClientId = previousAuthDetails.find(
-            (el) => el.verifier === ssoClient.type
-          )?.clientId;
+      authDetails.forEach((authDetail) => {
+        authSignerMakeTxValue[0].push(authDetail.verifier);
+        authSignerMakeTxValue[1].push(authDetail.clientId);
+      });
+      await signerMakeTx("setClientIds", authSignerMakeTxValue);
 
-          if (clientId && previousClientId !== clientId) {
-            loadingMessage.value = `Updating ${ssoClient.type} client id in smart contract...`;
-            console.log(
-              "Making tx for adding " + ssoClient.type + " client id"
-            );
-            const txResponse = await signerMakeTx({
-              ...getTxRequestProps(),
-              method: ssoClient.method,
-              value: [clientId],
-            });
-            console.log(
-              "Tx added for " + ssoClient.type + " client id",
-              txResponse
-            );
-          }
-        } catch (e) {
-          console.error("Tx failed for " + ssoClient.type);
-          console.error(e);
-        }
-      }
-
-      try {
-        if (
-          config.storage_limit !== previousConfig.storage_limit ||
-          config.bandwidth_limit !== previousConfig.bandwidth_limit
-        ) {
-          loadingMessage.value = `Updating user limits in smart contract...`;
-          const userLimitTxResponse = await signerMakeTx({
-            ...getTxRequestProps(),
-            method: "setDefaultLimit",
-            value: [config.storage_limit, config.bandwidth_limit],
-          });
-          console.log("Tx added for default limit", userLimitTxResponse);
-        }
-      } catch (e) {
-        console.error("Tx failed for default limit");
-        console.error(e);
-      }
-
-      try {
-        if (store.getters.appName !== previousConfig.appName) {
-          loadingMessage.value = `Updating app name in smart contract...`;
-          const appNameTxResponse = await signerMakeTx({
-            ...getTxRequestProps(),
-            method: "setAppName",
-            value: [store.getters.appName],
-          });
-          console.log("Tx added for app name", appNameTxResponse);
-        }
-      } catch (e) {
-        console.error("Tx failed for app name");
-        console.error(e);
+      if (
+        config.storage_limit !== previousConfig.storage_limit ||
+        config.bandwidth_limit !== previousConfig.bandwidth_limit
+      ) {
+        loadingMessage.value = `Updating user limits in smart contract...`;
+        await signerMakeTx("setDefaultLimit", [
+          config.storage_limit,
+          config.bandwidth_limit,
+        ]);
       }
 
       loading.value = false;
       router.push("/");
-    }
-
-    function getTxRequestProps() {
-      return {
-        privateKey: store.getters.keys.privateKey,
-        appAddress: store.getters.smartContractAddress,
-        rpc: store.getters["test/rpc"],
-        gateway: getEnvApi(),
-        forwarderAddress: store.getters["test/forwarder"],
-        accessToken: store.getters.accessToken,
-      };
     }
 
     function onFooterCancel() {
