--- src/utils/signer.js
@@ -1,1 +1,0 @@
-export const signerMakeTx = window.signerMakeTx;
